do not load partials as theme_assets
Sprockets handles loading the partials into the parent css.scss or
js.coffee files. Trying to load the partials into a ThemeAsset causes an error using Sprockets
find_asset helper as the partial is not an asset. The partial is just a
file the is imported into asset handled by Sprockets.